+++ packages/moonstone/Icon/tests/Icon-specs.js
@@ -15,6 +15,17 @@ describe('Icon Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
+	it('should return the correct Unicode value when provided \'star\' hex value', function () {
+		const icon = mount(
+			<Icon>0x0F0028</Icon>
+		);
+
+		const expected = 983080; // decimal converted charCode of character
+		const actual = icon.text().codePointAt();
+
+		expect(actual).to.equal(expected);
+	});
+
 	it('should return the correct Unicode value when provided HTML entity as hex value', function () {
 		const icon = mount(
 			<Icon>&#x02605;</Icon>
@@ -26,12 +37,21 @@ describe('Icon Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
+	it('should return the correct Unicode value when provided Unicode reference', function () {
+		const icon = mount(
+			<Icon>\u0F0028</Icon>
+		);
+
+		const expected = 983080; // decimal converted charCode of Unicode 'star' character
+		const actual = icon.text().codePointAt();
+
+		expect(actual).to.equal(expected);
+	});
+
 	it('should merge author styles with src', function () {
 		const src = 'images/icon.png';
 		const icon = shallow(
-			<Icon style={{color: 'green'}}>
-				{src}
-			</Icon>
+			<Icon src={src} style={{color: 'green'}} />
 		);
 
 		const expected = {