+++ packages/moonstone/CHANGELOG.md
@@ -12,13 +12,12 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Changed
 
-- `children` property of `moonstone/Icon` and `moonstone/IconButton` to support either a string name of an icon or a string or object representing the URI to the icon image. Also removed support for hex strings (`0x2704` or `\u2704`) in favor of escape sequences (`&#x2704`).
+- `children` property of `moonstone/Icon` and `moonstone/IconButton` to support either a string name of an icon or a string or object representing the URI to the icon image.
 - the `checked` property to `selected` for consistency across the whole framework. This allows better interoperability when switching between various components.  Affects the following: `CheckboxItem`, `RadioItem`, `SelectableItem`, `Switch`, `SwitchItem`, and `ToggleItem`. Additionally, these now use ItemOverlay to position and handle their Icons.
 - the implementation of `moonstone/GridListImageItem` so that a placeholder image displays while loading the image, and the caption and subcaption support marqueeing.
 
 ### Removed
 
-- `src` property from `moonstone/Icon` and `moonstone/IconButton` with support moving into the `children` property
 - LESS mixins that belong in `@enact/ui`, so that only moonstone-specific mixins are contained in
 this module. When authoring components and importing mixins, only the local mixins need to be
 imported, as they already import the general mixins.